refactor: next batch of changes
